refactor: move set-resolution helpers from _core.py to _defaults.py
