+++ src/envbool/_core.py
@@ -20,7 +20,7 @@ import logging
 from collections.abc import Iterable
 
 from envbool._config import _get_config
-from envbool._defaults import DEFAULT_FALSY, DEFAULT_TRUTHY
+from envbool._defaults import DEFAULT_FALSY, DEFAULT_TRUTHY, _apply_replace_or_extend
 from envbool.exceptions import InvalidBoolValueError
 
 # Module-level logger -- attributed to "envbool._core" so callers can filter it
@@ -138,41 +138,6 @@ def to_bool(
 
 
 # Private API
-
-
-def _normalize_set(values: Iterable[str]) -> frozenset[str]:
-    """Strip and lowercase values so they match to_bool()'s normalized input."""
-    return frozenset(v.strip().lower() for v in values)
-
-
-def _apply_replace_or_extend(
-    base: frozenset[str],
-    replace: Iterable[str] | None,
-    extend: Iterable[str] | None,
-) -> frozenset[str]:
-    """Resolve a value set using replace/extend/fall-back-to-base precedence.
-
-    Shared by _resolve() (call-site truthy/falsy args) and _config.py's
-    _parse_config() (TOML truthy/falsy keys) since both layer their inputs
-    on top of a base set using the same ruff select/extend-select pattern:
-        replace -- full replacement; caller owns the entire set
-        extend  -- additive; merges on top of base
-        neither -- use base as-is
-    replace takes precedence over extend; both cannot apply at once.
-
-    Args:
-        base: The starting set to fall back to or extend.
-        replace: If not None, fully replaces base (normalized).
-        extend: If not None and replace is None, merged on top of base.
-
-    Returns:
-        The resolved, normalized frozenset.
-    """
-    if replace is not None:
-        return _normalize_set(replace)
-    if extend is not None:
-        return base | _normalize_set(extend)
-    return base
 
 
 def _resolve(
